+++ src/main/java/com/travelservice/domain/product/entity/ProductImage.java
@@ -24,7 +24,7 @@ public class ProductImage {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "image_id")
-	private Integer imageId;
+	private Long imageId;
 
 	@Column(name = "image_url")
 	private String imageUrl;
